contracts

+++ src/main/java/com/example/allin/contract/ContractService.java
@@ -3,6 +3,7 @@ package com.example.allin.contract;
 import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -22,7 +23,11 @@ public class ContractService {
   }
 
   public ResponseEntity<Contract> getContractById(final String id) {
-    return ResponseEntity.status(HttpStatus.OK).body(contractRepo.findById(Integer.parseInt(id)).get());
+    Optional<Contract> contractOptional = contractRepo.findById(Integer.parseInt(id));
+    if (contractOptional.isEmpty()) {
+      return ResponseEntity.status(HttpStatus.NOT_FOUND).body(null);
+    }
+    return ResponseEntity.status(HttpStatus.OK).body(contractOptional.get());
   }
 
   public ResponseEntity<Contract> createContract(final Contract contract) {
@@ -30,12 +35,21 @@ public class ContractService {
   }
 
   public ResponseEntity<Contract> updateContract(final String id, final Contract contract) {
+    Optional<Contract> contractOptional = contractRepo.findById(Integer.parseInt(id));
+    if (contractOptional.isEmpty()) {
+      return ResponseEntity.status(HttpStatus.NOT_FOUND).body(null);
+    }
     return ResponseEntity.status(HttpStatus.OK).body(contractRepo.save(contract));
   }
 
-  public ResponseEntity<String> deleteContract(final String id) {
+  public ResponseEntity<Contract> deleteContract(final String id) {
+    Integer contract_id = Integer.parseInt(id);
+    Optional<Contract> contractOptional = contractRepo.findById(contract_id);
+    if (contractOptional.isEmpty()) {
+      return ResponseEntity.status(HttpStatus.NOT_FOUND).body(null);
+    }
     contractRepo.deleteById(Integer.parseInt(id));
-    return ResponseEntity.status(HttpStatus.NO_CONTENT).body("Contract deleted");
+    return ResponseEntity.status(HttpStatus.OK).body(contractOptional.get());
   }
 
 }
